requirements.txt
lightgbm==2.3.1
neptune-client==0.4.119
neptune-contrib==0.23.3
numpy==1.19.0
scikit-learn==0.23.1
scikit-plot==0.3.7
